# web/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FCC Workstation API",
    "description": "Generated, do not edit. Regenerate with .\\.venv\\Scripts\\python.exe -m frame_tools.report_api --write-openapi web\\src\\openapi.json",
    "version": "1.0.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/fields": {
      "get": {
        "summary": "Fields",
        "operationId": "fields_api_fields_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/report": {
      "get": {
        "summary": "Report",
        "operationId": "report_api_report_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Report"
                }
              }
            }
          }
        }
      }
    },
    "/api/fields/{field_id}/preview": {
      "post": {
        "summary": "Field Preview",
        "operationId": "field_preview_api_fields__field_id__preview_post",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Field Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/fields/{field_id}/value": {
      "post": {
        "summary": "Field Value",
        "operationId": "field_value_api_fields__field_id__value_post",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Field Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValueWriteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValueWriteResponse"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaleRevisionResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CheckModel": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "warn",
              "fail"
            ],
            "title": "Status"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "status",
          "name",
          "detail"
        ],
        "title": "CheckModel"
      },
      "FieldInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "min": {
            "type": "number",
            "title": "Min"
          },
          "max": {
            "type": "number",
            "title": "Max"
          },
          "file": {
            "type": "string",
            "title": "File"
          },
          "line": {
            "type": "integer",
            "title": "Line"
          },
          "current_value": {
            "title": "Current Value"
          },
          "status": {
            "type": "string",
            "enum": [
              "measured",
              "todo"
            ],
            "title": "Status"
          },
          "measurement_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measurement Label"
          },
          "group": {
            "type": "string",
            "title": "Group"
          }
        },
        "type": "object",
        "required": [
          "id",
          "question",
          "unit",
          "min",
          "max",
          "file",
          "line",
          "current_value",
          "status",
          "measurement_label",
          "group"
        ],
        "title": "FieldInfo"
      },
      "FieldsResponse": {
        "properties": {
          "revision": {
            "type": "string",
            "title": "Revision"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FieldInfo"
            },
            "type": "array",
            "title": "Fields"
          }
        },
        "type": "object",
        "required": [
          "revision",
          "fields"
        ],
        "title": "FieldsResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HeadlineItem": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Value"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          }
        },
        "type": "object",
        "required": [
          "label",
          "value",
          "unit"
        ],
        "title": "HeadlineItem"
      },
      "HealthResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "project_root": {
            "type": "string",
            "title": "Project Root"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "project_root"
        ],
        "title": "HealthResponse"
      },
      "PreviewRequest": {
        "properties": {
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              }
            ],
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "value"
        ],
        "title": "PreviewRequest"
      },
      "PreviewResponse": {
        "properties": {
          "diff": {
            "type": "string",
            "title": "Diff"
          }
        },
        "type": "object",
        "required": [
          "diff"
        ],
        "title": "PreviewResponse"
      },
      "Report": {
        "properties": {
          "headline": {
            "items": {
              "$ref": "#/components/schemas/HeadlineItem"
            },
            "type": "array",
            "title": "Headline"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/CheckModel"
            },
            "type": "array",
            "title": "Checks"
          }
        },
        "type": "object",
        "required": [
          "headline",
          "checks"
        ],
        "title": "Report"
      },
      "StaleRevisionResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "current": {
            "$ref": "#/components/schemas/FieldsResponse"
          }
        },
        "type": "object",
        "required": [
          "detail",
          "current"
        ],
        "title": "StaleRevisionResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValueWriteRequest": {
        "properties": {
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              }
            ],
            "title": "Value"
          },
          "revision": {
            "type": "string",
            "title": "Revision"
          }
        },
        "type": "object",
        "required": [
          "value",
          "revision"
        ],
        "title": "ValueWriteRequest"
      },
      "ValueWriteResponse": {
        "properties": {
          "result": {
            "$ref": "#/components/schemas/WriteResultModel"
          },
          "report": {
            "$ref": "#/components/schemas/Report"
          },
          "revision": {
            "type": "string",
            "title": "Revision"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "result",
          "report",
          "revision",
          "warnings"
        ],
        "title": "ValueWriteResponse"
      },
      "WriteResultModel": {
        "properties": {
          "file": {
            "type": "string",
            "title": "File"
          },
          "line_number": {
            "type": "integer",
            "title": "Line Number"
          },
          "old_text": {
            "type": "string",
            "title": "Old Text"
          },
          "new_text": {
            "type": "string",
            "title": "New Text"
          },
          "checklist_ticked": {
            "type": "boolean",
            "title": "Checklist Ticked"
          }
        },
        "type": "object",
        "required": [
          "file",
          "line_number",
          "old_text",
          "new_text",
          "checklist_ticked"
        ],
        "title": "WriteResultModel"
      }
    }
  },
  "x-generated": "Generated, do not edit. Regenerate with .\\.venv\\Scripts\\python.exe -m frame_tools.report_api --write-openapi web\\src\\openapi.json"
}